requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
tensorflow>=2.16.0
Pillow==10.0.0
numpy==1.24.3
werkzeug==2.3.7
python-dotenv==1.0.0